refactor(editor): Rename redirection core to match the composable-name convention (no-changelog)
Per Operator convention (file name = composable name), rename the base
composable off the `.core` suffix:

- `usePageRedirectionHelper.core.ts` -> `useBasePageRedirectionHelper.ts`, and
  its export `usePageRedirectionHelper` -> `useBasePageRedirectionHelper`.
- The app-facing wrapper keeps the name `usePageRedirectionHelper` and binds the
  guard; its import of the base is de-aliased.
- Tests follow the rename.

No behavior change and no change to the ~40 call sites, which still import
`usePageRedirectionHelper` from the same path.
